+++ src/index.ts
@@ -62,5 +62,8 @@ export * from './components/common/dom/changeInputData';
 export * from './components/SelectTree';
 export * from './components/Tree';
 export * from './components/Table';
+export * from './components/InputIconButton';
+export * from './components/GroupActionsPane';
 
 export { FontsVTBGroup } from './components/FontsVTBGroup';
+export { DropdownProvider } from './components/DropdownProvider';